scripts/requirements.txt
absl-py==0.13.0
aiohttp==3.8.5
aiosignal==1.3.1
alphafold-colabfold==2.3.5
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
biopython==1.81
biotite==0.37.0
boto3==1.28.26
botocore==1.31.26
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
chex==0.1.5
click==8.1.6
colabfold==1.2.0
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.36
deepspeed==0.5.9
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-haiku==0.0.5
dm-tree==0.1.8
docker==6.1.3
einops==0.6.1
fair-esm==2.0.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.5.4
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.8.0
hdbscan==0.8.30
hjson==3.1.0
huggingface-hub==0.16.4
idna==3.4
iminuit==1.5.4
immutabledict==2.2.5
importlib-metadata==6.7.0
jax==0.3.25
jaxlib==0.3.25
jmespath==1.0.1
jmp==0.0.4
joblib==1.3.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
lightning-utilities==0.9.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.1.3
ml-collections==0.1.1
msgpack==1.0.5
multidict==6.0.4
networkx==2.6.3
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkg_resources==0.0.0
polyleven==0.8
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
py3Dmol==1.8.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytorch-transformers==1.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
six==1.16.0
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.4
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tmscoring==0.4
tokenizers==0.13.3
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.65.0
transformers==4.25.1
triton==1.0.0
typing_extensions==4.7.1
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
